Path planning
Introduces path planning with cost based behaviour and trajectory generation.

The path planning is implemented by PathPlanner.

The flow could be simplified as ...

```
def onMessage (localization_input):
  val path_prev = localization_input.path;
  val cars = parse(localization_input.sensor_fusion)

  //// PathPlanner
  // Behavior
  val successor_state = get_successor_state(path_prev, cars)
  // Trajectory
  path = get_trajectory(successor_state, path_prev)

```

Provides a path on cartesian x & y axis for the given state and observations. Internally, it consist of 2 distinct parts...

1. Behaviour, for suggesting successor state based on current state and observations
2. Trajectory generation, for executing suggested successor state in the highway environment

Provides best successor state for prior state and observations based on the cost.

The cost function takes into account both speed and cars/objects in the surrounding lanes. It provides the combination of the `speed_cost` and the `lane_cost`

For both cost functions, position of closest surrounding cars are labelled as follows

|S distance	| Left lane		| Centre lane	| Right lane
|:-:		| :-:			| :-:			| :-:			|
|s > 0		| Left - Ahead	| Ahead			| Right - Ahead	|
|s ~= 0		| Left - Nearby	| Car			| Right - Ahead	|
|s < 0		| Left - Behind	|				| Right - Behind	|

The cost function aims to penalise the slow traffic and crowded lanes

Generates a trajectory for a given prior path and successor state.

The trajectory x & y point values are calculated using a spline. To generate a spline, we sample known/deducted path points and estimated points towards the goal. Known path points taken from 2 last known/estimated positions at the end of prior path, and target points consist of 3 intermediate step towards the goal that we aim to move into in the future.

For left and right turns, the goal/target point horizon is adjusted based on speed. i.e., faster the car, longer the horizon. This helps us to reduce jerk and have a smoother transition

The code is tested on docker image. Therefore, it is recommended to use following  steps

Build base image using `base.sh`

Build and run project image using `run.sh`